+++ src/components/App/App.tsx
@@ -1,6 +1,6 @@
 import { Route, Routes } from 'react-router-dom';
 import { MainPage } from '../../pages/Main/MainPage';
-import { Page404 } from '../../pages/Main/Page404/Page404';
+import { Page404 } from '../../pages/Page404/Page404';
 
 export function App(): JSX.Element {
   return (
