+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import { store } from './store/store';
+import { store } from './redux/store';
 import { ErrorBoundary } from './components/ErrorBoundary/ErrorBoundary';
 import { App } from './components/App/App';
 import './index.css';
